Register new users PR (#4)
* Update ID type fromt Integer to Long

* Implements asimmetric password decryption with Private Key, pending password encoding

* EnableJpaAuditing

* Implements profile to local and remote database server

* Implement Docker configuration to deploy on koyeb for testing

* Update CORS configuration

* Update Security Configuration

* Add dockercompose

* Generate docker infrastructure local - dev - prod

* Update docker compose using unique docker hub

* Update per_verification_code varchar size to 3000

* Add JWT and Resend Mail dependencies and config

* Implements mail configuration

* Implements registration and verify token process. It changes the endpoint's path.

* Implements Controller Advice using by Custom Exception Manager

* Improve throw exceptions on Register Method and implements Valid annotation to control RegistrationDto with new dependencies.

* Implements freemarker to send register mail with custom template

* Improve validation annotations

* Add Enum for Roles and implements in all classes

* Implements Record to get Verification code on registration process.

* Implements generic message response and new codes.

* Start documentation with OpenApi and Swagger

* Implement custom validator for blank spaces at init or end

* Implement unit test with JUnit and Mockito

* Implements countries endpoint

+++ src/main/java/ec/com/expensys/persistence/repository/ExpPersonCategoryRepository.java
@@ -3,6 +3,6 @@ package ec.com.expensys.persistence.repository;
 import ec.com.expensys.persistence.entity.ExpPersonCategory;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface ExpPersonCategoryRepository extends ListCrudRepository<ExpPersonCategory, Integer> {
+public interface ExpPersonCategoryRepository extends ListCrudRepository<ExpPersonCategory, Long> {
 
 }
